pyproject.toml
[tool.poetry]
name = "supernova"
version = "0.1.0"
description = "AI-powered development assistant within the terminal"
authors = ["Nikhil Laturi <nikhil.j2se@gmail.com>"]
readme = "README.md"
packages = [{include = "supernova"}]

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.7"
prompt-toolkit = "^3.0.41"
litellm = "^1.7.0"
pydantic = "^2.4.2"
pyyaml = "^6.0.1"
python-dotenv = "^1.0.0"
gitpython = "^3.1.40"
pathspec = "^0.11.2"
rich = "^13.6.0"
sqlite3 = {version = "^0.0.0", optional = true}  # Uses system SQLite

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pytest-cov = "^4.1.0"
black = "^23.3.0"
mypy = "^1.3.0"
isort = "^5.12.0"
pylint = "^2.17.4"

[tool.poetry.scripts]
supernova = "supernova.cli.main:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pylint.messages_control]
disable = ["C0111", "C0103", "C0330", "C0326", "W0511", "R0903"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "--cov=supernova --cov-report=term-missing --cov-fail-under=80"
python_functions = "test_*" 